requirements.txt
torch>=2.7.0
torch-geometric>=2.6.0
torch-scatter>=2.1.0
torch-sparse>=0.6.18
torchvision>=0.22.0
numpy>=2.2.0
pandas>=2.3.0
scikit-learn>=1.7.0
matplotlib>=3.10.0
umap-learn>=0.5.0
